Se ordenaron los archivos
Se creo una nueva aplicacion de django para almacenar todo el formulario
