+++ bug_tracking/bug_report/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class BugReportConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'bug_report'
