+++ bug_tracking/bug_report/templates/bug_report/confirmacion_reporte.html
@@ -1,0 +1,8 @@
+{% extends 'home/base.html' %}
+{% block content %}
+<section class='home'>
+  <h1 class="text">Confirmación de reporte de bug</h1>
+  <p class="text">Tu reporte ha sido enviado con éxito. ¡Gracias por tu ayuda en mejorar nuestro sistema!</p>
+  <a href="{% url 'home:principal' %}" class="text">Volver a la página principal</a>
+</section>
+  {% endblock %}
