+++ bug_tracking/bug_report/templates/bug_report/reporte_bug.html
@@ -1,7 +1,7 @@
-{% extends './layouts/base.html' %}
+{% extends 'home/base.html' %}
 {% block content %}
 <section class='home'>
-  <h1>Reportar un bug</h1>
+  <h1 class="text">Reportar un bug</h1>
   <form method="post">
     <!-- csrf es para proteccion contra crost site request forgeries -->
     {% csrf_token %}
