+++ bug_tracking/bug_tracking/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
     'home',
     'database',
     'buglist',
+    'bug_report',
 ]
 
 MIDDLEWARE = [
