+++ bug_tracking/bug_tracking/urls.py
@@ -21,4 +21,5 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include("home.urls")),
     path('buglist', include("buglist.urls")),
+    path('report', include("bug_report.urls")),
 ]
