+++ bug_tracking/buglist/templates/buglist/index.html
@@ -1,4 +1,4 @@
-{% extends 'home/layouts/base.html' %}
+{% extends 'home/base.html' %}
 {% block content %}
     <section class="home">
         <div class="text">Lista de bugs</div>
