+++ bug_tracking/buglist/urls.py
@@ -4,6 +4,6 @@ from . import views
 
 app_name = "buglist"
 urlpatterns = [
-    path('', views.index, name= 'bugs_list'),
-    path('back',views.buglist_home, name='coming_home'),
+    path('', views.index, name= 'bug_list'),
+    # path('back',views.buglist_home, name='coming_home'),
 ]
