+++ bug_tracking/home/templates/home/base.html
@@ -49,14 +49,14 @@
                     </li>
 
                     <li class="nav-link">
-                        <a href="{% url 'home:buglist' %}">
+                        <a href="{% url 'buglist:bug_list' %}">
                             <i class='bx bx-list-ul icon'></i>
                             <span class="text nav-text">Lista de bugs</span>
                         </a>
                     </li>
 
                     <li class="nav-link">
-                        <a href="reportar_bug">
+                        <a href="{% url 'report:reportar_bug' %}">
                             <i class='bx bx-bell icon' ></i>
                             <span class="text nav-text">Reportar bug</span>
                         </a>
