--- bug_tracking/home/templates/home/confirmacion_reporte.html
@@ -1,8 +1,0 @@
-{% extends './layouts/base.html' %}
-{% block content %}
-<section class='home'>
-  <h1>Confirmación de reporte de bug</h1>
-  <p>Tu reporte ha sido enviado con éxito. ¡Gracias por tu ayuda en mejorar nuestro sistema!</p>
-  <a href="{% url 'home:principal' %}">Volver a la página principal</a>
-</section>
-  {% endblock %}
