--- bug_tracking/home/templates/home/index.html
@@ -1,36 +1,0 @@
-{% extends './layouts/base.html' %}
-{% block content %}
-  <section class="home">
-       <div class="text">No se que texto iba aca xd</div>
-       <table class="table">
-           <thead>
-               <tr>
-                 <th>Rank</th>
-                 <th>Name</th>
-                 <th>Points</th>
-                 <th>Team</th>
-               </tr>
-             </thead>
-             <tbody>
-               <tr>
-                 <td>1</td>
-                 <td>Domenic</td>
-                 <td>88,110</td>
-                 <td>dcode</td>
-               </tr>
-               <tr>
-                 <td>2</td>
-                 <td>Sally</td>
-                 <td>72,400</td>
-                 <td>Students</td>
-               </tr>
-               <tr>
-                 <td>3</td>
-                 <td>Nick</td>
-                 <td>52,300</td>
-                 <td>dcode</td>
-               </tr>
-             </tbody>
-       </table>
-   </section>
-{% endblock %}
